[CMake] DYNAMIC_GRAPH_PLUGINDIR is now relative
For now, the plugins can only be loaded from python, so we just need the
wrap.so to be able to find its plugin.so. For this, there is no need for
a unique hard-coded absolute path to provide an RPATH to
libdynamic-graph.so.

And actually, that RPATH was not even set.